Perform pull member up on serialize method
