+++ bvm/ballerina-runtime/src/main/java/org/ballerinalang/jvm/values/XMLSequence.java
@@ -17,13 +17,11 @@
 package org.ballerinalang.jvm.values;
 
 import org.ballerinalang.jvm.BallerinaErrors;
-import org.ballerinalang.jvm.BallerinaXMLSerializer;
 import org.ballerinalang.jvm.XMLNodeType;
 import org.ballerinalang.jvm.types.BArrayType;
 import org.ballerinalang.jvm.types.BTypes;
 import org.ballerinalang.jvm.util.BLangConstants;
 import org.ballerinalang.jvm.util.exceptions.BallerinaErrorReasons;
-import org.ballerinalang.jvm.util.exceptions.BallerinaException;
 import org.ballerinalang.jvm.values.api.BMap;
 import org.ballerinalang.jvm.values.api.BString;
 import org.ballerinalang.jvm.values.api.BXML;
@@ -31,8 +29,6 @@ import org.ballerinalang.jvm.values.freeze.FreezeUtils;
 import org.ballerinalang.jvm.values.freeze.State;
 import org.ballerinalang.jvm.values.freeze.Status;
 
-import java.io.IOException;
-import java.io.OutputStream;
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
@@ -388,27 +384,6 @@ public final class XMLSequence extends XMLValue {
         }
 
         return new XMLSequence(descendants);
-    }
-
-    // Methods from Datasource impl
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public void serialize(OutputStream outputStream) {
-        if (outputStream instanceof BallerinaXMLSerializer) {
-            ((BallerinaXMLSerializer) outputStream).write(this);
-        } else {
-            BallerinaXMLSerializer ballerinaXMLSerializer = new BallerinaXMLSerializer(outputStream);
-            ballerinaXMLSerializer.write(this);
-            try {
-                ballerinaXMLSerializer.flush();
-                ballerinaXMLSerializer.close();
-            } catch (IOException e) {
-                throw new BallerinaException(e);
-            }
-        }
     }
 
     /**
